refactor: 은행 선택 및 자산연동 페이지로 이동
