+++ src/components/BankIcon.vue
